code base splitted into components including css
