+++ src/components/header/header.css
@@ -1,0 +1,34 @@
+.header {
+    grid-area: header;
+  
+    display: grid;
+    grid-template-columns: 180px 1fr;
+  
+    border-bottom: solid 1px #00000020;
+}
+  
+.header > .left-menu {
+    display: grid;
+    grid-auto-flow: column;
+    align-items: center;
+    justify-content: left;
+}
+  
+.header > .left-menu .menu-icon {
+    width: 25px;
+    height: 25px;
+    margin: 10px;
+}
+  
+.header > .search {
+    display: grid;
+    padding: 6px;
+}
+  
+.header > .search .search-bar {
+    border: none;
+    border-radius: 5px;
+    background-color: rgb(240, 240, 240);
+    padding: 0 10px;
+    font-size: 1rem;
+}
