+++ src/components/main/subMainComponents/ArticleHeader.jsx
@@ -1,6 +1,6 @@
-import rateStarButton from './assets/icons/rate-star-button.png'
-import backArrow from './assets/icons/back-arrow.png'
-import rubbishButton from './assets/icons/rubbish-bin-delete-button.png'
+import rateStarButton from '../../../assets/icons/rate-star-button.png'
+import backArrow from '../../../assets/icons/back-arrow.png'
+import rubbishButton from '../../../assets/icons/rubbish-bin-delete-button.png'
 
 
 function articleHeader() {
