+++ src/components/main/subMainComponents/ContentNav.jsx
@@ -1,6 +1,6 @@
-import downloadButton from './assets/icons/download-button.png'
-import rubbishButton from './assets/icons/rubbish-bin-delete-button.png'
-import backArrow from './assets/icons/back-arrow.png'
+import downloadButton from '../../../assets/icons/download-Button.png'
+import rubbishButton from '../../../assets/icons/rubbish-bin-delete-button.png'
+import backArrow from '../../../assets/icons/back-arrow.png'
 
 function contentNav() {
     return (
