+++ src/components/nav/nav.css
@@ -1,0 +1,29 @@
+.inbox-list > .left-menu {
+    grid-area: left-menu;
+    padding: 10px;
+    padding-left: 0;
+  }
+  
+.inbox-list > .left-menu .inbox-list {
+    display: grid;
+    grid-gap: 10px;
+}
+  
+.inbox-list > .left-menu .inbox-list .item {
+    padding: 5px 10px;
+    display: grid;
+    grid-auto-flow: column;
+    justify-content: space-between;
+    align-items: center;
+  
+    border-top-right-radius: 20px;
+    border-bottom-right-radius: 20px;
+}
+
+.left-menu .inbox-list .item.toggle {
+    background-color: transparent;
+}
+
+.left-menu .inbox-list .item.active {
+    background-color: #ff000050;
+}
